A simple class for storing UTF-8 strings

The values are stored as a character() vector. On construction, the enc2utf8() function is called on the input. Subsetting and concatenation operations on an object of this class return an object of this class again. Calls to Encoding<-() are not intercepted.

utf8(x = character())

is_utf8(x)

as_utf8(x, ...)

# S3 method for default
as_utf8(x, ...)

# S3 method for `NULL`
as_utf8(x, ...)

# S3 method for character
as_utf8(x, ...)

# S3 method for utf8
as_utf8(x, ...)

# S3 method for utf8
as.character(x, ...)

# S3 method for utf8
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse = " ")
)

# S3 method for utf8
format(x, ...)

# S3 method for utf8
print(x, ...)

Arguments

x

A vector

...

Arguments passed on to further methods.

row.names

NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional

logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

nm

Name of column in new data frame

Examples

utf8(letters)
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"
utf8("ä")
#> [1] "ä"
utf8(iconv("ä", to = "latin1"))
#> [1] "ä"